+++ src/widget/sideBar/createPostModal/addPostSection/AddPostSection.module.scss
