+++ src/widget/sideBar/createPostModal/closeCreatePostModal/CloseCreatePostModal.module.scss
