+++ src/widget/sideBar/createPostModal/cropAndScaleSection/CropAndScaleSection.module.scss
@@ -1,5 +1,4 @@
-// @import '@/myApp/styles/mixins';
-@import '../../../myApp/styles/mixins';
+@import '@/myApp/styles/mixins';
 
 .wrapper {
   // position: relative;
