+++ src/widget/sideBar/createPostModal/cropAndScaleSection/CropAndScaleSection.tsx~
@@ -4,16 +4,17 @@ import Cropper, { Area, Point } from 'react-easy-crop';
 import ArrowIosBack from '@/../public/assets/components/ArrowIosBack';
 import ArrowIosForward from '@/../public/assets/components/ArrowIosForward';
 import { useAppDispatch, useAppSelector } from '@/shared/lib';
-import { AddNewImgPanel } from '@/widget/sideBar/addNewImgPanel/AddNewImgPanel';
-import { AspectRatioVals } from '@/widget/sideBar/createPostModal/createPost';
-import { createPostActions, createPostSelectors } from '@/widget/sideBar/createPostModal/createPostSlice';
-import { CropAndScalePanel, aspectRatios } from '@/widget/sideBar/cropAndScalePanel/CropAndScalePanel';
 import { clsx } from 'clsx';
 import { Swiper as SwiperProps } from 'swiper';
 import { Navigation } from 'swiper/modules';
 import { Swiper, SwiperSlide } from 'swiper/react';
 
 import s from './CropAndScaleSection.module.scss';
+
+import { CropAndScalePanel } from '../cropAndScaleSection/cropAndScalePanel/CropAndScalePanel';
+import { AspectRatioVals, aspectRatios } from '../lib/createPostConstants';
+import { createPostActions, createPostSelectors } from '../service/createPostSlice';
+import { AddNewImgPanel } from './addNewImgPanel/AddNewImgPanel';
 
 type CropAndScaleSectionType = {
   errorMessage?: null | string;
@@ -23,7 +24,7 @@ type CropAndScaleSectionType = {
 /**
  * Компонента СЕКЦИИ для обрезания и масштабирования картинки
  * @param {(e: ChangeEvent<HTMLInputElement>) => void} props.onSelectFile - обязательный параметр, функция для добавления изображения, которая передается в AddNewImgPanel
- * * Сохраняем индекс активной картинки карусели Swiper (используем тут) и передаем в AddNewImgPanel (чтобы в DropDown можно было переключаться по картинкам тоже)
+ * @description * Сохраняем индекс активной картинки карусели Swiper (используем тут) и передаем в AddNewImgPanel (чтобы в DropDown можно было переключаться по картинкам тоже)
  * * В Swiper отображаем Cropper компонентой от ReactEasyCrop котрая принимает crop область, aspect, zoom и функции для их изменения
  * * AddNewImgPanel - компонент для добавления новых изображений
  * * CropAndScalePanel - компонент для масштабирования и обрезания
