+++ src/widget/sideBar/createPostModal/cropAndScaleSection/addNewImgPanel/AddNewImgPanel.module.scss
