+++ src/widget/sideBar/createPostModal/cropAndScaleSection/cropAndScalePanel/CropAndScalePanel.module.scss
