+++ src/widget/sideBar/createPostModal/cropAndScaleSection/cropAndScalePanel/CropAndScalePanel.tsx~
@@ -1,27 +1,22 @@
+import type { AspectRatioVals } from '../../lib/createPostConstants';
+
 import React from 'react';
 
 import Expand from '@/../public/assets/components/Expand';
 import IconHorizontalRectangle from '@/../public/assets/components/IconHorizontalRectangle';
 import IconSquare from '@/../public/assets/components/IconSquare';
 import IconVerticalRectangle from '@/../public/assets/components/IconVerticalRectangle';
+import Maximize from '@/../public/assets/components/Maximize';
+import MaximizeOutline from '@/../public/assets/components/MaximizeOutline';
 import Picture from '@/../public/assets/components/PictureOutline';
 import { useAppSelector } from '@/shared/lib';
 import { CustomDropdownItem, CustomDropdownWrapper, Slider, Typography } from '@/shared/ui';
-import { AspectRatioVals } from '@/widget/sideBar/createPostModal/createPost';
-import { createPostSelectors } from '@/widget/sideBar/createPostModal/createPostSlice';
 import { clsx } from 'clsx';
 
 import s from './CropAndScalePanel.module.scss';
 
-import Maximize from '../../../../public/assets/components/Maximize';
-import MaximizeOutline from '../../../../public/assets/components/MaximizeOutline';
-
-export const aspectRatios: AspectRatioVals[] = [
-  { text: 'Original', value: 1 / 1 },
-  { text: '1:1', value: 1 / 1 },
-  { text: '4:5', value: 4 / 5 },
-  { text: '16:9', value: 16 / 9 }
-];
+import { aspectRatios } from '../../lib/createPostConstants';
+import { createPostSelectors } from '../../service/createPostSlice';
 
 type PropsCropAndScale = {
   id: string;
