+++ src/widget/sideBar/createPostModal/cropAndScaleSection/cropAndScalePanel/CropAndScalePost.module.scss~
