+++ src/widget/sideBar/createPostModal/filtersSection/FiltersSection.module.scss
