+++ src/widget/sideBar/createPostModal/lib/cropImage.ts
