+++ src/widget/sideBar/createPostModal/lib/navigateBtnLogic.ts~
