+++ src/widget/sideBar/createPostModal/lib/useModalTitle.ts~
