+++ src/widget/sideBar/createPostModal/lib/useSelectFilesAndShowError.ts~
