+++ src/widget/sideBar/createPostModal/publicationSection/PublicationSection.module.scss
