+++ src/widget/sideBar/createPostModal/publicationSection/PublicationSection.tsx~
@@ -12,8 +12,8 @@ import {
 import { publicApi } from '@/shared/api/public/publicApi';
 import { useAppDispatch, useAppSelector, useCustomToast } from '@/shared/lib';
 import { FormTextfield, FormTextfieldArea, Typography } from '@/shared/ui';
-import { createPostSelectors } from '@/widget/sideBar/createPostModal/createPostSlice';
-import { useRefreshPostCreationData } from '@/widget/sideBar/lib/useRefreshPostCreationData';
+import { createPostSelectors } from '../service/createPostSlice';
+import { useRefreshPostCreationData } from '../hooks/useRefreshPostCreationData';
 import { zodResolver } from '@hookform/resolvers/zod';
 import Image from 'next/image';
 import { z } from 'zod';
